[server] moved tests/ to test/ and WIP on the other stuff
